drop time dependency

+++ executors/src/timeconstants.rs
@@ -7,15 +7,15 @@
 // except according to those terms.
 
 #[allow(dead_code)]
-pub const NS_PER_S: u64 = 1_000_000_000;
+pub const NS_PER_S: u32 = 1_000_000_000;
 #[allow(dead_code)]
-pub const US_PER_S: u64 = 1_000_000;
+pub const US_PER_S: u32 = 1_000_000;
 #[allow(dead_code)]
-pub const MS_PER_S: u64 = 1_000;
+pub const MS_PER_S: u32 = 1_000;
 
 #[allow(dead_code)]
-pub const NS_PER_MS: u64 = 1_000_000;
+pub const NS_PER_MS: u32 = 1_000_000;
 #[allow(dead_code)]
-pub const US_PER_MS: u64 = 1_000;
+pub const US_PER_MS: u32 = 1_000;
 #[allow(dead_code)]
-pub const NS_PER_US: u64 = 1_000;
+pub const NS_PER_US: u32 = 1_000;
